Setup minimal reproduction of using `puppeteer`
This is as slim as we can get it (barring stylistic choices).

The brunt of this is that we seem to have to download the `chromium`
binary ourselves. Otherwise, the one that is downloaded doesn't seem to
work for some reason.

The other part is that we seem to have to wrap the `nodejs_binary` in an
`npm_package_bin` in order to pass the `chromium` binary to it properly.
It seems to lose the `chromium` binary if it's passed as a `data`
attribute.

Hopefully it's clear what all the rest of the stuff is doing.